+++ tests/unit/stub/SauceLabs.js
@@ -1,0 +1,11 @@
+/*
+* Copyright (c) 2014, Yahoo! Inc.  All rights reserved.
+* Copyrights licensed under the New BSD License.
+* See the accompanying LICENSE file for terms.
+*/
+
+function SauceLabs() {
+}
+
+module.exports = SauceLabs;
+